requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
bandit==1.6.2
colorama==0.4.3
flake8==3.8.3
gitdb==4.0.5
GitPython==3.1.8
iniconfig==1.0.1
mccabe==0.6.1
more-itertools==8.5.0
packaging==20.4
pbr==5.5.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PySDL2==0.9.7
pytest==6.0.1
PyYAML==5.3.1
six==1.15.0
smmap==3.0.4
stevedore==3.2.1
toml==0.10.1
